requirements.txt
pillow
streamlit
torch
torchvision